backend/requirements.txt
Django==5.0.4
djangorestframework==3.15.1
drf_spectacular==0.27.2
